pyproject.toml
[tool.poetry]
name = "project-farm"
version = "0.1.0"
description = ""
authors = ["danquimby <danquimby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.2.1"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.8.1"
loguru = "^0.7.3"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
orjson = "^3.10.15"
censys = "^2.2.17"
censys-platform = "^0.5.1"
alembic = "^1.13.2"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
pytest-asyncio = "^1.0.0"
pytest = "^8.4.1"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "^4.3.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
